Issue #4287 - Move getUriLastPathSegment() to URIUtil
